Refactor: 退出時の処理を分離
